Выполнено задание #6 Тесные связи
