+++ js/game-2.js
@@ -1,4 +1,6 @@
 import createElementFromTemplate from './create-element-from-template';
+import renderElement from './render-element';
+import game3Element from './game-3';
 
 let template = `<header class="header">
   <div class="header__back">
@@ -46,4 +48,12 @@ let template = `<header class="header">
 </div>`;
 
 const moduleElement = createElementFromTemplate(template);
+
+moduleElement.querySelector('.game').onclick = (e) => {
+  if (e.target.parentElement.classList.contains('game__answer')) {
+    renderElement(game3Element);
+  }
+};
+
+
 export default moduleElement;
