+++ js/game-3.js
@@ -1,4 +1,6 @@
 import createElementFromTemplate from './create-element-from-template';
+import renderElement from './render-element';
+import statsElement from './stats';
 
 let template = `<header class="header">
   <div class="header__back">
@@ -44,4 +46,11 @@ let template = `<header class="header">
 </div>`;
 
 const moduleElement = createElementFromTemplate(template);
+
+moduleElement.querySelector('.game').onclick = (e) => {
+  if (e.target.classList.contains('game__option')) {
+    renderElement(statsElement);
+  }
+};
+
 export default moduleElement;
