+++ js/intro.js
@@ -1,4 +1,6 @@
 import createElementFromTemplate from './create-element-from-template';
+import renderElement from './render-element';
+import greetingElement from './greeting';
 
 let template = `<div id="intro" class="intro">
   <h1 class="intro__asterisk">*</h1>
@@ -7,4 +9,7 @@ let template = `<div id="intro" class="intro">
 </div>`;
 
 const moduleElement = createElementFromTemplate(template);
+
+moduleElement.querySelector('.intro__asterisk').onclick = () => renderElement(greetingElement);
+
 export default moduleElement;
