+++ js/render-element.js
@@ -1,0 +1,6 @@
+let mainElement = document.getElementById('main');
+
+export default (element) => {
+  mainElement.innerHTML = '';
+  mainElement.appendChild(element);
+};
